Support var() resolution in paint() arguments
Our var() support in the CSS parser relies on the fact that
property parser functions do not succeed parsing when var() exists
in the token sequence [1]. However, the parser for the paint()
currently allows *any* token sequence, also token sequences with
var(). This results in var() functions not working at all inside
paint() function arguments.

This CL changes the paint() parser to use CSSVariableParser to
check if any var() references are found, and if so, fail parsing,
thus ultimately producing a CSSVariableReferenceValue (as it should)
via [1].

Also added a DCHECK in CSSPaintValue::ParseInputArguments which
verifies that we don't have any unresolved references. Although this
DCHECK would pass on current 'main' as well, since the paint() parser
basically lied and set this flag to 'false' unconditionally.

Note that this CL only affects code behind the CSSPaintAPIArguments
flag (status:experimental).

[1] See CSSPropertyParser::ParseValueStart

Change-Id: Id4ef6a7774ffbbb9fe38fb6a5506f744b40ec02e