URLPattern: Use default options for cannot-be-a-base pathnames.
Standard URLs have well structured pathnames with `/` characters.  We
therefore use liburlpattern options that make `/` characters behave as
prefixes and separators when operating on pathnames.

Non-standard (or "cannot be a base URL") URLs do not treat `/` specially
in pathnames, though.  Therefore we should use default liburlpatterion
options in this case to avoid incorrectly treating a `/` character as
a prefix, etc.

This CL includes a test that fails without this change.

Bug: 1234865
Change-Id: I7a4e3216fb63b8c34c27662bb0bcc236669f947c